Move `ConcurrentLruMap` to inner class of `CrossRegionS3Client`.
