+++ spring-cloud-aws-s3-parent/spring-cloud-aws-s3-cross-region-client/src/test/java/io/awspring/cloud/s3/crossregion/ConcurrentLruMapTests.java
@@ -15,6 +15,7 @@
  */
 package io.awspring.cloud.s3.crossregion;
 
+import static io.awspring.cloud.s3.crossregion.CrossRegionS3Client.ConcurrentLruMap;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.concurrent.ExecutorService;
